agent/pyproject.toml
[project]
name = "wardenhub-agent"
version = "0.1.0"
description = "WardHub security auditing agent"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0",
    "prometheus-client>=0.20",
    "httpx>=0.27",
    "typer>=0.12",
    "structlog>=24.0",
]

[project.scripts]
wardenhub-agent = "wardenhub_agent.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short -m 'not integration'"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.14",
    "ruff>=0.4",
]
